Fixed get_mean function to take care of nan as well
