+++ tests/src/chanjo2/meta/test_handle_report_contents.py
@@ -4,7 +4,6 @@ from sqlalchemy.orm import sessionmaker
 
 from chanjo2.demo import DEMO_COVERAGE_QUERY_FORM
 from chanjo2.meta.handle_report_contents import (
-    get_mean,
     get_missing_genes_from_db,
     get_report_data,
 )
@@ -28,26 +27,6 @@ REPORT_EXPECTED_EXTRA_KEYS = [
     "completeness_thresholds",
     "samples",
 ]
-
-
-def test_get_mean_floats():
-    """Test invoking the get_mean function with a list of floating point numbers."""
-
-    # GIVEN a list of float numbers
-    value_list = [12.81, 34.72, 22.53]
-
-    # get_mean should return a float
-    assert isinstance(get_mean(float_list=value_list), float)
-
-
-def test_get_mean_inf():
-    """Test invoking the get_mean function with a list of numbers that contains a non-number."""
-
-    # GIVEN a list of floats and an inf
-    value_list = [float("inf"), 7, 45.22]
-
-    # get_mean should return a string
-    assert get_mean(float_list=value_list) == "inf"
 
 
 def test_get_missing_genes_from_db(
